client/pyproject.toml
[project]
name = "couch-client"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "websockets>=14.0",
    "pyyaml>=6.0",
    "sounddevice>=0.4.6",
    "numpy>=1.26.0",
]

[tool.uv]
package = false
